// src/modules/pppoe-auditoria/domain/enums/pppoe-auditoria-enums.ts
export enum OrigenOperacionPppoe {
  OPERADOR = 'OPERADOR',
  SISTEMA = 'SISTEMA',
  COBRANZA_AUTOMATICA = 'COBRANZA_AUTOMATICA',
}

export enum AccionAuditoriaPppoe {
  /*
   * Homologación de perfiles
   */
  PERFIL_HOMOLOGADO = 'PERFIL_HOMOLOGADO',
  PERFIL_ACTUALIZADO = 'PERFIL_ACTUALIZADO',
  PERFIL_ACTIVADO = 'PERFIL_ACTIVADO',
  PERFIL_DESACTIVADO = 'PERFIL_DESACTIVADO',

  /*
   * Prealta y credenciales
   */
  PREALTA_CREADA = 'PREALTA_CREADA',
  PREALTA_CANCELADA = 'PREALTA_CANCELADA',
  CONTRASENA_GENERADA = 'CONTRASENA_GENERADA',

  /*
   * Creación del secret
   */
  CREACION_AUTORIZADA = 'CREACION_AUTORIZADA',
  CREACION_CANCELADA = 'CREACION_CANCELADA',
  SECRET_CREADO = 'SECRET_CREADO',

  /*
   * Activación y reactivación
   */
  SECRET_HABILITADO = 'SECRET_HABILITADO',
  SERVICIO_ACTIVADO = 'SERVICIO_ACTIVADO',
  SERVICIO_REACTIVADO = 'SERVICIO_REACTIVADO',

  /*
   * Suspensión
   */
  SECRET_DESHABILITADO = 'SECRET_DESHABILITADO',
  SERVICIO_SUSPENDIDO = 'SERVICIO_SUSPENDIDO',
  SESION_ACTIVA_REMOVIDA = 'SESION_ACTIVA_REMOVIDA',

  /*
   * Desinstalación
   */
  DESINSTALACION_AUTORIZADA = 'DESINSTALACION_AUTORIZADA',
  DESINSTALACION_RECHAZADA = 'DESINSTALACION_RECHAZADA',
  SECRET_ELIMINADO = 'SECRET_ELIMINADO',

  /*
   * Reautenticación
   */
  REAUTENTICACION_EXITOSA = 'REAUTENTICACION_EXITOSA',
  REAUTENTICACION_FALLIDA = 'REAUTENTICACION_FALLIDA',

  /*
   * Operaciones técnicas
   */
  OPERACION_INICIADA = 'OPERACION_INICIADA',
  OPERACION_EXITOSA = 'OPERACION_EXITOSA',
  OPERACION_FALLIDA = 'OPERACION_FALLIDA',
  OPERACION_PARCIAL = 'OPERACION_PARCIAL',
  OPERACION_CANCELADA = 'OPERACION_CANCELADA',
  OPERACION_REINTENTADA = 'OPERACION_REINTENTADA',
  OPERACION_CREADA = 'OPERACION_CREADA',
  OPERACION_RECUPERADA = 'OPERACION_RECUPERADA',

  /*
   * Sincronización con MikroTik
   */
  SINCRONIZACION_EXITOSA = 'SINCRONIZACION_EXITOSA',
  SINCRONIZACION_FALLIDA = 'SINCRONIZACION_FALLIDA',
  DESINCRONIZACION_DETECTADA = 'DESINCRONIZACION_DETECTADA',

  /*
   * Documentos y hojas operativas
   */
  HOJA_VISUALIZADA = 'HOJA_VISUALIZADA',
  HOJA_GENERADA = 'HOJA_GENERADA',
}